split the huge shotgun utils file into several chunks.
