+++ python/tank/util/shotgun/__init__.py
@@ -1,0 +1,29 @@
+# Copyright (c) 2017 Shotgun Software Inc.
+# 
+# CONFIDENTIAL AND PROPRIETARY
+# 
+# This work is provided "AS IS" and subject to the Shotgun Pipeline Toolkit 
+# Source Code License included in this distribution package. See LICENSE.
+# By accessing, using, copying or modifying this work you indicate your 
+# agreement to the Shotgun Pipeline Toolkit Source Code License. All rights 
+# not expressly granted therein are reserved by Shotgun Software Inc.
+
+
+from .connection import \
+    get_project_name_studio_hook_location, \
+    get_associated_sg_base_url, \
+    get_associated_sg_config_data, \
+    get_deferred_sg_connection, \
+    get_sg_connection, \
+    create_sg_connection
+
+from .publish_util import \
+    get_entity_type_display_name, \
+    find_publish, \
+    create_event_log_entry, \
+    get_published_file_entity_type
+
+from .publish_creation import register_publish
+from .publish_resolve import resolve_publish_path
+from .download import download_url, download_and_unpack_attachment
+
