+++ tests/python/tank_test/tank_test_base.py
@@ -280,8 +280,8 @@ class TankTestBase(unittest.TestCase):
         # Mock this so that authentication manager works even tough we are not in a config.
         # If we don't mock it than the path cache calling get_current_user will fail.
         self._mock_return_value(
-            "tank.util.shotgun.get_associated_sg_config_data",
-            {"host": "https://somewhere.shotguntudio.com"}
+            "tank.util.shotgun.connection.get_associated_sg_config_data",
+            {"host": "https://somewhere.shotgunstudio.com"}
         )
 
         # define entity for test project
@@ -357,6 +357,8 @@ class TankTestBase(unittest.TestCase):
         # fake a version response from the server
         self.mockgun.server_info = {"version": (7, 0, 0)}
 
+        self._mock_return_value("tank.util.shotgun.connection.get_associated_sg_base_url", "http://unit_test_mock_sg")
+        self._mock_return_value("tank.util.shotgun.connection.create_sg_connection", self.mockgun)
         self._mock_return_value("tank.util.shotgun.get_associated_sg_base_url", "http://unit_test_mock_sg")
         self._mock_return_value("tank.util.shotgun.create_sg_connection", self.mockgun)
 
@@ -405,7 +407,7 @@ class TankTestBase(unittest.TestCase):
                 os.remove(path_cache_file)
 
             # clear global shotgun accessor
-            tank.util.shotgun._g_sg_cached_connections = threading.local()
+            tank.util.shotgun.connection._g_sg_cached_connections = threading.local()
 
             # get rid of init cache
             if os.path.exists(pipelineconfig_factory._get_cache_location()):
